site_t (instead of int) is used for the size of work-memory in CCR12. Some minor fix in comments etc included.
--HG--
extra : convert_revision : svn%3A0fb12191-db4b-0410-8866-e78deaee71e5/trunk%409246

+++ mpqc/src/lib/chemistry/qc/ccr12/ccr12_info.h
@@ -143,7 +143,7 @@ class CCR12_Info : virtual public RefCount {
     std::vector<unsigned int> mosyma_,mosymb_;
     std::vector<unsigned int> cabssyma_,cabssymb_;
     const Ref<R12IntEvalInfo>& r12evalinfo_;
-    int workmemsize_;
+    size_t workmemsize_;
 
     // orbital data after sorted
     std::vector<double> orbital_evl_sorted_;
